+++ MummyDiner/state/GameState.h
@@ -9,26 +9,15 @@
 #ifndef __MummyDiner__GameState__
 #define __MummyDiner__GameState__
 
-#include <stdio.h>
-#include <iostream>
+#include "../util/FramesPerSecond.h"
 
-#include <SFML/Graphics.hpp>
+#include "../gui/ButtonClass.h" // includes text class.. also the debug class
+#include "../gui/BackgroundClass.h"
 
-#include "util/DebugMode.h"
-#include "util/FramesPerSecond.h"
-#include "util/TextClass.h"
-#include "util/ButtonClass.h"
-#include "util/BackgroundClass.h"
-
-#include "sprite/SpriteClass.h"
-#include "sprite/Waitress.h"
-#include "sprite/Customer.h"
-#include "sprite/Chef.h"
-
-#define DEBUG_MODE 1
-
-#define QUICKSAND_REGULAR_FONT "fonts/Quicksand-Regular.ttf"
-#define LATO_LIGHT_FONT "fonts/Lato-Light.ttf"
+// each of these classes includes the SpriteClass class and debug class
+#include "../sprite/Waitress.h"
+#include "../sprite/Customer.h"
+#include "../sprite/Chef.h"
 
 #define MOUSE_X event.mouseMove.x
 #define MOUSE_Y event.mouseMove.y
@@ -42,11 +31,6 @@ enum state {
 	SETTINGS,
 	LEVEL
 };
-
-const int SCREEN_W = 640;
-const int SCREEN_H = 480;
-
-using namespace sf;
 
 class GameState {
 	public:
@@ -69,7 +53,7 @@ class GameState {
 		Event event;
 		Color color;
 	
-		DebugMode _debug;
+		Utility _debug;
 	
 		int _mouseXPos;
 		int _mouseYPos;