+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 ###CHANGELOG
 
+**12 March 2015 - v0.4.0-alpha**
+- Added smoke image
+- Waitress can now collide with objects
+- There is now frames per second and capping
+- Restructured and renamed some project files
+- Rewrote main menu
+- Rewrote sprite classes
+- Rewrote level classes
+- Improved button class
+- Fixed a bug where button text is not rendered
+- Fixed a bug where placement of text on a button is not centered
+- Fixed a bug that caused debug text to not render
+- Fixed a bug where FPS does not change in value
+- Fixed a bug where button width bounds for mouse click is incorrect
+- Fixed a bug which prevents sprite from moving
+
 **4 March 2015 - v0.3.0-alpha**
 - Rewrote debug mode
 - Refactored button class